requirements.txt
coloredlogs
discord.py
python-dotenv